requirements.txt

pandas
numpy
setuptools
requests
google-cloud-storage
pyyaml
matplotlib
tensorflow==2.21.0
wordcloud
comet-ml
dvc
dvc-gs
flask
flask-cors
joblib
python-dotenv
simplejson>=3.19.0